requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
jinja2
pyyaml
sqlalchemy